requirements.txt
asgiref==3.7.2
distlib==0.3.8
Django==4.2.5
django-ckeditor==6.7.0
django-js-asset==2.2.0
django-model-utils==4.5.1
djangorestframework==3.15.1
django-import-export==3.3.6
filelock==3.15.4
pillow==10.3.0
platformdirs==4.2.2
psycopg2==2.9.9
sqlparse==0.4.4
tzdata==2024.1
